+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Support for ipv6 in `oci_core_vcn`, `oci_core_subnet` and `oci_load_balancer` resources.
 - Support for ipv6 in `oci_core_virtual_circuit` resources.
 
+### Fixed
+- Fixed the diff for `options` arguments order in `oci_core_dhcp_options` resource [Issue #829](https://github.com/terraform-providers/terraform-provider-oci/issues/829)
+
 ## 3.36.0 (July 31, 2019)
 
 ### Added